Move findVocabularyProperties
